remove factory
