--- agential/cog/critic/factory.py
@@ -1,266 +1,0 @@
-"""CRITIC prompts and fewshot examples selector."""
-
-from typing import Any, Dict
-
-from agential.cog.base.factory import BaseFactory
-from agential.cog.constants import BENCHMARK_FEWSHOTS, Benchmarks, FewShotType
-from agential.cog.critic.prompts import (
-    AMBIGNQ_FEWSHOT_EXAMPLES_CRITIC,
-    CRITIC_CRITIQUE_INSTRUCTION_AMBIGNQ,
-    CRITIC_CRITIQUE_INSTRUCTION_FEVER,
-    CRITIC_CRITIQUE_INSTRUCTION_GSM8K,
-    CRITIC_CRITIQUE_INSTRUCTION_HOTPOTQA,
-    CRITIC_CRITIQUE_INSTRUCTION_HUMANEVAL,
-    CRITIC_CRITIQUE_INSTRUCTION_MBPP,
-    CRITIC_CRITIQUE_INSTRUCTION_SVAMP,
-    CRITIC_CRITIQUE_INSTRUCTION_TABMWP,
-    CRITIC_CRITIQUE_INSTRUCTION_TRIVIAQA,
-    CRITIC_CRITIQUE_NO_TOOL_INSTRUCTION_GSM8K,
-    CRITIC_CRITIQUE_NO_TOOL_INSTRUCTION_HUMANEVAL,
-    CRITIC_CRITIQUE_NO_TOOL_INSTRUCTION_MBPP,
-    CRITIC_CRITIQUE_NO_TOOL_INSTRUCTION_SVAMP,
-    CRITIC_CRITIQUE_NO_TOOL_INSTRUCTION_TABMWP,
-    CRITIC_INSTRUCTION_AMBIGNQ,
-    CRITIC_INSTRUCTION_FEVER,
-    CRITIC_INSTRUCTION_HOTPOTQA,
-    CRITIC_INSTRUCTION_TRIVIAQA,
-    CRITIC_POT_INSTRUCTION_GSM8K,
-    CRITIC_POT_INSTRUCTION_HUMANEVAL,
-    CRITIC_POT_INSTRUCTION_MBPP,
-    CRITIC_POT_INSTRUCTION_SVAMP,
-    CRITIC_POT_INSTRUCTION_TABMWP,
-    FEVER_FEWSHOT_EXAMPLES_CRITIC,
-    GSM8K_FEWSHOT_EXAMPLES_CRITIC,
-    GSM8K_FEWSHOT_EXAMPLES_CRITIC_NO_TOOL,
-    HOTPOTQA_FEWSHOT_EXAMPLES_CRITIC,
-    HUMANEVAL_FEWSHOT_EXAMPLES_CRITIC,
-    HUMANEVAL_FEWSHOT_EXAMPLES_CRITIC_NO_TOOL,
-    MBPP_FEWSHOT_EXAMPLES_CRITIC,
-    MBPP_FEWSHOT_EXAMPLES_CRITIC_NO_TOOL,
-    SVAMP_FEWSHOT_EXAMPLES_CRITIC,
-    SVAMP_FEWSHOT_EXAMPLES_CRITIC_NO_TOOL,
-    TABMWP_FEWSHOT_EXAMPLES_CRITIC,
-    TABMWP_FEWSHOT_EXAMPLES_CRITIC_NO_TOOL,
-    TRIVIAQA_FEWSHOT_EXAMPLES_CRITIC,
-)
-from agential.cog.critic.strategies.base import CriticBaseStrategy
-from agential.cog.critic.strategies.code import (
-    CriticHEvalCodeStrategy,
-    CriticMBPPCodeStrategy,
-)
-from agential.cog.critic.strategies.math import (
-    CriticGSM8KStrategy,
-    CriticSVAMPStrategy,
-    CriticTabMWPStrategy,
-)
-from agential.cog.critic.strategies.qa import (
-    CriticAmbigNQStrategy,
-    CriticFEVERStrategy,
-    CriticHotQAStrategy,
-    CriticTriviaQAStrategy,
-)
-
-CRITIC_BENCHMARK_FEWSHOTS = {
-    Benchmarks.HOTPOTQA: [FewShotType.COT, FewShotType.DIRECT, FewShotType.REACT],
-    Benchmarks.FEVER: [FewShotType.COT, FewShotType.DIRECT, FewShotType.REACT],
-    Benchmarks.TRIVIAQA: [FewShotType.COT, FewShotType.DIRECT, FewShotType.REACT],
-    Benchmarks.AMBIGNQ: [FewShotType.COT, FewShotType.DIRECT, FewShotType.REACT],
-    Benchmarks.GSM8K: [FewShotType.POT],
-    Benchmarks.SVAMP: [FewShotType.POT],
-    Benchmarks.TABMWP: [FewShotType.POT],
-    Benchmarks.HUMANEVAL: [FewShotType.POT],
-    Benchmarks.MBPP: [FewShotType.POT],
-}
-
-
-CRITIC_PROMPTS = {
-    Benchmarks.HOTPOTQA: {
-        "prompt": CRITIC_INSTRUCTION_HOTPOTQA,
-        "critique_prompt": CRITIC_CRITIQUE_INSTRUCTION_HOTPOTQA,
-        "critique_prompt_no_tool": CRITIC_CRITIQUE_INSTRUCTION_HOTPOTQA,
-    },
-    Benchmarks.FEVER: {
-        "prompt": CRITIC_INSTRUCTION_FEVER,
-        "critique_prompt": CRITIC_CRITIQUE_INSTRUCTION_FEVER,
-        "critique_prompt_no_tool": CRITIC_CRITIQUE_INSTRUCTION_FEVER,
-    },
-    Benchmarks.TRIVIAQA: {
-        "prompt": CRITIC_INSTRUCTION_TRIVIAQA,
-        "critique_prompt": CRITIC_CRITIQUE_INSTRUCTION_TRIVIAQA,
-        "critique_prompt_no_tool": CRITIC_CRITIQUE_INSTRUCTION_TRIVIAQA,
-    },
-    Benchmarks.AMBIGNQ: {
-        "prompt": CRITIC_INSTRUCTION_AMBIGNQ,
-        "critique_prompt": CRITIC_CRITIQUE_INSTRUCTION_AMBIGNQ,
-        "critique_prompt_no_tool": CRITIC_CRITIQUE_INSTRUCTION_AMBIGNQ,
-    },
-    Benchmarks.GSM8K: {
-        "prompt": CRITIC_POT_INSTRUCTION_GSM8K,
-        "critique_prompt": CRITIC_CRITIQUE_INSTRUCTION_GSM8K,
-        "critique_prompt_no_tool": CRITIC_CRITIQUE_NO_TOOL_INSTRUCTION_GSM8K,
-    },
-    Benchmarks.SVAMP: {
-        "prompt": CRITIC_POT_INSTRUCTION_SVAMP,
-        "critique_prompt": CRITIC_CRITIQUE_INSTRUCTION_SVAMP,
-        "critique_prompt_no_tool": CRITIC_CRITIQUE_NO_TOOL_INSTRUCTION_SVAMP,
-    },
-    Benchmarks.TABMWP: {
-        "prompt": CRITIC_POT_INSTRUCTION_TABMWP,
-        "critique_prompt": CRITIC_CRITIQUE_INSTRUCTION_TABMWP,
-        "critique_prompt_no_tool": CRITIC_CRITIQUE_NO_TOOL_INSTRUCTION_TABMWP,
-    },
-    Benchmarks.HUMANEVAL: {
-        "prompt": CRITIC_POT_INSTRUCTION_HUMANEVAL,
-        "critique_prompt": CRITIC_CRITIQUE_INSTRUCTION_HUMANEVAL,
-        "critique_prompt_no_tool": CRITIC_CRITIQUE_NO_TOOL_INSTRUCTION_HUMANEVAL,
-    },
-    Benchmarks.MBPP: {
-        "prompt": CRITIC_POT_INSTRUCTION_MBPP,
-        "critique_prompt": CRITIC_CRITIQUE_INSTRUCTION_MBPP,
-        "critique_prompt_no_tool": CRITIC_CRITIQUE_NO_TOOL_INSTRUCTION_MBPP,
-    },
-}
-
-CRITIC_FEWSHOTS = {
-    Benchmarks.HOTPOTQA: {
-        "critique_examples": HOTPOTQA_FEWSHOT_EXAMPLES_CRITIC,
-        "critique_examples_no_tool": HOTPOTQA_FEWSHOT_EXAMPLES_CRITIC,
-    },
-    Benchmarks.FEVER: {
-        "critique_examples": FEVER_FEWSHOT_EXAMPLES_CRITIC,
-        "critique_examples_no_tool": FEVER_FEWSHOT_EXAMPLES_CRITIC,
-    },
-    Benchmarks.TRIVIAQA: {
-        "critique_examples": TRIVIAQA_FEWSHOT_EXAMPLES_CRITIC,
-        "critique_examples_no_tool": TRIVIAQA_FEWSHOT_EXAMPLES_CRITIC,
-    },
-    Benchmarks.AMBIGNQ: {
-        "critique_examples": AMBIGNQ_FEWSHOT_EXAMPLES_CRITIC,
-        "critique_examples_no_tool": AMBIGNQ_FEWSHOT_EXAMPLES_CRITIC,
-    },
-    Benchmarks.GSM8K: {
-        "critique_examples": GSM8K_FEWSHOT_EXAMPLES_CRITIC,
-        "critique_examples_no_tool": GSM8K_FEWSHOT_EXAMPLES_CRITIC_NO_TOOL,
-    },
-    Benchmarks.SVAMP: {
-        "critique_examples": SVAMP_FEWSHOT_EXAMPLES_CRITIC,
-        "critique_examples_no_tool": SVAMP_FEWSHOT_EXAMPLES_CRITIC_NO_TOOL,
-    },
-    Benchmarks.TABMWP: {
-        "critique_examples": TABMWP_FEWSHOT_EXAMPLES_CRITIC,
-        "critique_examples_no_tool": TABMWP_FEWSHOT_EXAMPLES_CRITIC_NO_TOOL,
-    },
-    Benchmarks.HUMANEVAL: {
-        "critique_examples": HUMANEVAL_FEWSHOT_EXAMPLES_CRITIC,
-        "critique_examples_no_tool": HUMANEVAL_FEWSHOT_EXAMPLES_CRITIC_NO_TOOL,
-    },
-    Benchmarks.MBPP: {
-        "critique_examples": MBPP_FEWSHOT_EXAMPLES_CRITIC,
-        "critique_examples_no_tool": MBPP_FEWSHOT_EXAMPLES_CRITIC_NO_TOOL,
-    },
-}
-
-CRITIC_STRATEGIES = {
-    Benchmarks.HOTPOTQA: CriticHotQAStrategy,
-    Benchmarks.FEVER: CriticFEVERStrategy,
-    Benchmarks.TRIVIAQA: CriticTriviaQAStrategy,
-    Benchmarks.AMBIGNQ: CriticAmbigNQStrategy,
-    Benchmarks.GSM8K: CriticGSM8KStrategy,
-    Benchmarks.SVAMP: CriticSVAMPStrategy,
-    Benchmarks.TABMWP: CriticTabMWPStrategy,
-    Benchmarks.HUMANEVAL: CriticHEvalCodeStrategy,
-    Benchmarks.MBPP: CriticMBPPCodeStrategy,
-}
-
-
-class CriticFactory(BaseFactory):
-    """A factory class for creating instances of Critic strategies and selecting prompts and few-shot examples."""
-
-    @staticmethod
-    def get_fewshots(
-        benchmark: str, fewshot_type: str, **kwargs: Any
-    ) -> Dict[str, str]:
-        """Retrieve few-shot examples based on the benchmark.
-
-        Args:
-            benchmark (str): The benchmark name.
-            fewshot_type (str): The benchmark few-shot type.
-            **kwargs (Any): Additional arguments.
-
-        Returns:
-            Dict[str, str]: A dictionary of few-shot examples.
-        """
-        if (
-            benchmark not in CRITIC_FEWSHOTS
-            or benchmark not in CRITIC_BENCHMARK_FEWSHOTS
-        ):
-            raise ValueError(f"Benchmark '{benchmark}' few-shots not found for Critic.")
-
-        if fewshot_type not in CRITIC_BENCHMARK_FEWSHOTS[benchmark]:
-            raise ValueError(
-                f"Benchmark '{benchmark}' few-shot type not supported for Critic."
-            )
-
-        benchmark_fewshots = BENCHMARK_FEWSHOTS[benchmark][fewshot_type]
-
-        use_tool = kwargs.get("use_tool")
-        if use_tool is None:
-            raise ValueError("`use_tool` not specified.")
-
-        if use_tool:
-            return {
-                "examples": benchmark_fewshots,
-                "critique_examples": CRITIC_FEWSHOTS[benchmark]["critique_examples"],
-            }
-        return {
-            "examples": benchmark_fewshots,
-            "critique_examples": CRITIC_FEWSHOTS[benchmark][
-                "critique_examples_no_tool"
-            ],
-        }
-
-    @staticmethod
-    def get_prompts(benchmark: str, **kwargs: Any) -> Dict[str, str]:
-        """Retrieve the prompt instruction based on the benchmark.
-
-        Args:
-            benchmark (str): The benchmark name.
-            **kwargs (Any): Additional arguments.
-
-        Returns:
-            Dict[str, str]: The prompt instructions.
-        """
-        if benchmark not in CRITIC_PROMPTS:
-            raise ValueError(f"Benchmark '{benchmark}' prompt not found for Critic.")
-
-        use_tool = kwargs.get("use_tool")
-        if use_tool is None:
-            raise ValueError("`use_tool` not specified.")
-
-        if use_tool:
-            return {
-                "prompt": CRITIC_PROMPTS[benchmark]["prompt"],
-                "critique_prompt": CRITIC_PROMPTS[benchmark]["critique_prompt"],
-            }
-        return {
-            "prompt": CRITIC_PROMPTS[benchmark]["prompt"],
-            "critique_prompt": CRITIC_PROMPTS[benchmark]["critique_prompt_no_tool"],
-        }
-
-    @staticmethod
-    def get_strategy(benchmark: str, **kwargs: Any) -> CriticBaseStrategy:
-        """Returns an instance of the appropriate Critic strategy based on the provided benchmark.
-
-        Args:
-            benchmark (str): The benchmark name.
-            **kwargs (Any): Additional keyword arguments to pass to
-                the strategy's constructor.
-
-        Returns:
-            CriticBaseStrategy: An instance of the appropriate Critic strategy.
-        """
-        if benchmark not in CRITIC_STRATEGIES:
-            raise ValueError(f"Unsupported benchmark: {benchmark} for agent Critic")
-
-        strategy = CRITIC_STRATEGIES[benchmark]
-        return strategy(**kwargs)
